Native bridge: Delete `workdir` before linking
This is necessary, because the optimizer will produce a different number
of files depending on its configuration. The files produced by a
previous run, with a different configuration, may be conflicting with a
subsequent run.
